Security: bound skippable frame loops, guard FSE bit reads, widen metadata to u64
- Skippable frame skip loop: add MAX_SKIPPABLE_FRAMES=256 depth limit and
  bounds-check total_frame_size against remaining input to prevent DoS via
  crafted streams with thousands of skippable frames.
- FSE read_bits_from_buffer: reject num_bits > 57 (max safe with bit_off
  up to 7 inside a u64) and document caller responsibility for buffer length.
- NvcompMetadata.uncompressed_size: widen from u32 to u64 to support files
  >4GB without silent truncation. Also widen NvcompV5Metadata uncompressed_size
  and compressed_size to u64. Remove narrowing cast at metadata extraction.

+++ include/cuda_zstd_nvcomp.h
@@ -145,8 +145,8 @@ struct NvcompV5Metadata {
     u32 format_version;           // NVCOMP format version
     u32 library_version;          // Library version
     int compression_level;        // Compression level used
-    u32 uncompressed_size;        // Original size
-    u32 compressed_size;          // Compressed size
+    u64 uncompressed_size;        // Original size (u64 to support >4GB)
+    u64 compressed_size;          // Compressed size
     u32 num_chunks;               // Number of chunks
     u32 chunk_size;               // Size per chunk
     u32 dictionary_id;            // Dictionary ID (0 if none)
